requirements.txt
aiohttp==3.8.5 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
altair==5.0.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.10" and python_version < "4.0"
cohere==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
dataclasses-json==0.5.9 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
docx2txt==0.8 ; python_version >= "3.10" and python_version < "4.0"
faiss-cpu==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.3 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2023.9.0 ; python_version >= "3.10" and python_version < "4.0"
gitdb==4.0.10 ; python_version >= "3.10" and python_version < "4.0"
gitpython==3.1.32 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
huggingface-hub==0.16.4 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.18.4 ; python_version >= "3.10" and python_version < "4.0"
langchain==0.0.220 ; python_version >= "3.10" and python_version < "4.0"
langchainplus-sdk==0.0.20 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
marshmallow-enum==1.5.1 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.20.1 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
numexpr==2.8.4 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
openai==0.27.8 ; python_version >= "3.10" and python_version < "4.0"
openapi-schema-pydantic==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.23.4 ; python_version >= "3.10" and python_version < "4.0"
pyarrow; python_version >= "3.10" and python_version < "4.0"
pycryptodome==3.18.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.11 ; python_version >= "3.10" and python_version < "4.0"
pydeck==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.15.1 ; python_version >= "3.10" and python_version < "4.0"
pympler==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pymupdf==1.22.5 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.30.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2023.6.3 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rich==13.4.2 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.9.2 ; python_version >= "3.10" and python_version < "4.0"
safetensors==0.3.3 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.19 ; python_version >= "3.10" and python_version < "4.0"
streamlit==1.25.0 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.2 ; python_version >= "3.10" and python_version < "4.0"
tiktoken==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
tornado==6.3.2 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
transformers==4.33.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
tzlocal==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "4.0"
validators==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
watchdog==3.0.0 ; python_version >= "3.10" and python_version < "4.0" and platform_system != "Darwin"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.16.2 ; python_version >= "3.10" and python_version < "4.0"
-e ./
